Put `cerberus` back
Sadly, going from one service to two, especially when the second one is a
watchdog, requires some kind of init system.  This time around I decided to go
with `s6`.  I used runit before but that's in the testing repo for alpine and
this is fully packaged.

I'd use `nosh` but then I'd have to compile it all myself.